Adding Users to Groups with File

+++ src/main/java/parsso/idman/RepoImpls/GroupRepoImpl.java
@@ -2,6 +2,14 @@ package parsso.idman.RepoImpls;
 
 
 import net.minidev.json.JSONObject;
+import org.apache.poi.hssf.usermodel.HSSFSheet;
+import org.apache.poi.hssf.usermodel.HSSFWorkbook;
+import org.apache.poi.ss.usermodel.Cell;
+import org.apache.poi.ss.usermodel.DataFormatter;
+import org.apache.poi.ss.usermodel.Row;
+import org.apache.poi.ss.usermodel.Sheet;
+import org.apache.poi.xssf.usermodel.XSSFSheet;
+import org.apache.poi.xssf.usermodel.XSSFWorkbook;
 import org.json.simple.JSONArray;
 import org.json.simple.parser.ParseException;
 import org.slf4j.Logger;
@@ -16,6 +24,8 @@ import org.springframework.ldap.filter.AndFilter;
 import org.springframework.ldap.filter.EqualsFilter;
 import org.springframework.ldap.support.LdapNameBuilder;
 import org.springframework.stereotype.Service;
+import org.springframework.web.multipart.MultipartFile;
+import parsso.idman.Helpers.User.BuildDn;
 import parsso.idman.Models.Group;
 import parsso.idman.Models.SimpleUser;
 import parsso.idman.Models.User;
@@ -25,11 +35,11 @@ import parsso.idman.Repos.UserRepo;
 
 import javax.naming.Name;
 import javax.naming.NamingException;
-import javax.naming.directory.Attributes;
-import javax.naming.directory.BasicAttribute;
-import javax.naming.directory.BasicAttributes;
-import javax.naming.directory.SearchControls;
+import javax.naming.directory.*;
+import java.io.BufferedReader;
 import java.io.IOException;
+import java.io.InputStream;
+import java.io.InputStreamReader;
 import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.LinkedList;
@@ -39,16 +49,14 @@ import java.util.List;
 public class GroupRepoImpl implements GroupRepo {
 
     Logger logger = LoggerFactory.getLogger(Group.class);
-
+    @Autowired
+    BuildDn buildDnUser;
     @Value("${spring.ldap.base.dn}")
     private String BASE_DN;
-
     @Autowired
     private LdapTemplate ldapTemplate;
-
     @Autowired
     private UserRepo userRepo;
-
     @Autowired
     private ServiceRepo serviceRepo;
 
@@ -60,12 +68,40 @@ public class GroupRepoImpl implements GroupRepo {
             groups = retrieve();
         else {
             ArrayList jsonArray = (ArrayList) jsonObject.get("names");
+            for (Object string:jsonArray
+                 ) {
+                Group group = retrieveOu((String) string);
+                if (group != null)
+                    groups.add(group);
+                else
+                    continue;
+            }
             Iterator<String> iterator = jsonArray.iterator();
             while (iterator.hasNext()) {
                 Group group = retrieveOu(iterator.next());
                 if (group != null)
                     groups.add(group);
+                else
+                    continue;
+
+                DirContextOperations context;
+
+                String id = group.getId();
+
+                for (SimpleUser user : userRepo.retrieveGroupsUsers(id)) {
+                    if(user!=null && user.getMemberOf()!=null)
+                    for (String groupN : user.getMemberOf()) {
+                        if (groupN.equalsIgnoreCase(id)) {
+                            context = ldapTemplate.lookupContext(buildDnUser.buildDn(user.getUserId()));
+                            context.removeAttributeValue("ou", id);
+                            ldapTemplate.modifyAttributes(context);
+                        }
+
+                    }
+                }
             }
+
+
         }
 
         if (groups != null)
@@ -81,6 +117,8 @@ public class GroupRepoImpl implements GroupRepo {
 
             }
 
+
+
         logger.info("Removing groups ended");
         return HttpStatus.OK;
     }
@@ -94,7 +132,7 @@ public class GroupRepoImpl implements GroupRepo {
         List<Group> groups = retrieve();
 
         for (Group group : groups) {
-            if (group.getId().equals(uid))
+            if (group.getId().equalsIgnoreCase(uid))
                 return group;
         }
         return null;
@@ -113,6 +151,8 @@ public class GroupRepoImpl implements GroupRepo {
         }
         return groups;
     }
+
+
 
     private Attributes buildAttributes(String uid, Group group) {
 
@@ -155,9 +195,9 @@ public class GroupRepoImpl implements GroupRepo {
         SearchControls searchControls = new SearchControls();
         searchControls.setSearchScope(SearchControls.SUBTREE_SCOPE);
 
-        List<Group> gt = ldapTemplate.search(BASE_DN,null, new OUAttributeMapper());
+        List<Group> gt = ldapTemplate.search(BASE_DN, null, new OUAttributeMapper());
 
-        if (gt.size()!=0)
+        if (gt.size() != 0)
             return null;
 
         final AndFilter filter = new AndFilter();
@@ -174,9 +214,9 @@ public class GroupRepoImpl implements GroupRepo {
 
         List<Group> groups = retrieve();
 
-        for (Group group:groups)
+        for (Group group : groups)
             if (group.getId().equals(ou.getId()))
-                return  HttpStatus.FOUND;
+                return HttpStatus.FOUND;
 
         Name dn = buildDn(ou.getId());
         try {
@@ -197,46 +237,49 @@ public class GroupRepoImpl implements GroupRepo {
 
         List<Group> groups = retrieve();
 
-        for (Group group:groups)
-            if (!(id.equals(ou.getId()))&&group.getId().equals(ou.getId()))
-                return  HttpStatus.FOUND;
-
-        DirContextOperations context = ldapTemplate.lookupContext(dn);
+        for (Group group : groups)
+            if (!(id.equals(ou.getId())) && group.getId().equals(ou.getId()))
+                return HttpStatus.FOUND;
 
 
-        if (!(id.equals(ou.getId()))){
+        if (!(id.equals(ou.getId()))) {
 
             try {
                 ldapTemplate.unbind(dn);
                 create(ou);
+                DirContextOperations contextUser;
                 logger.info("Group " + ou.getId() + " updated successfully");
 
-                for (SimpleUser user:userRepo.retrieveGroupsUsers(id) ) {
-                    for (String group:user.getMemberOf()) {
-                        if (group.equals(id)){
-                            context.removeAttributeValue("ou", id);
-                            context.addAttributeValue("ou", ou.getId());
+                for (SimpleUser user : userRepo.retrieveGroupsUsers(id)) {
+                    for (String group : user.getMemberOf()) {
+                        if (group.equalsIgnoreCase(id)) {
+                            contextUser = ldapTemplate.lookupContext(buildDnUser.buildDn(user.getUserId()));
+                            contextUser.removeAttributeValue("ou", id);
+                            contextUser.addAttributeValue("ou", ou.getId());
+                            ldapTemplate.modifyAttributes(contextUser);
                         }
-
                     }
                 }
 
-                for (parsso.idman.Models.Service service: serviceRepo.listServicesWithGroups(id)) {
 
-                    //remove old id and add new id
-                    ((List<String>) ((JSONArray) service.getAccessStrategy().getRequiredAttributes().get("ou")).get(1)).remove(id);
-                    ((List<String>) ((JSONArray) service.getAccessStrategy().getRequiredAttributes().get("ou")).get(1)).add(ou.getId());
+                List<parsso.idman.Models.Service> services = serviceRepo.listServicesWithGroups(id);
+                if (services != null)
+                    for (parsso.idman.Models.Service service : services) {
 
-                    // delete old service
-                    org.json.simple.JSONObject jsonObject = new org.json.simple.JSONObject();
-                    jsonObject.put("names",((((JSONArray) service.getAccessStrategy().getRequiredAttributes().get("ou")).get(1))));
-                    serviceRepo.deleteServices(jsonObject);
+                        //remove old id and add new id
+                        ((List<String>) ((JSONArray) service.getAccessStrategy().getRequiredAttributes().get("ou")).get(1)).remove(id);
+                        ((List<String>) ((JSONArray) service.getAccessStrategy().getRequiredAttributes().get("ou")).get(1)).add(ou.getId());
 
-                    // create new service
+                        // delete old service
+                        org.json.simple.JSONObject jsonObject = new org.json.simple.JSONObject();
+                        jsonObject.put("names", ((((JSONArray) service.getAccessStrategy().getRequiredAttributes().get("ou")).get(1))));
+                        serviceRepo.deleteServices(jsonObject);
 
-                    serviceRepo.updateOuIdChange(service,service.getId(),service.getName(),id,ou.getId());
+                        // create new service
 
-                }
+                        serviceRepo.updateOuIdChange(service, service.getId(), service.getName(), id, ou.getId());
+
+                    }
 
                 logger.info("Group " + ou.getId() + " updated successfully");
                 return HttpStatus.OK;
@@ -251,7 +294,7 @@ public class GroupRepoImpl implements GroupRepo {
                 return HttpStatus.BAD_REQUEST;
             }
 
-        }else {
+        } else {
 
             try {
                 ldapTemplate.rebind(dn, null, buildAttributes(ou.getId(), ou));
@@ -264,8 +307,10 @@ public class GroupRepoImpl implements GroupRepo {
                 return HttpStatus.BAD_REQUEST;
             }
         }
-    }
 
+
+
+    }
